# internal/handlers/modelversion/create/modelversion.go
package modelversion

import (
	"encoding/json"
	"fmt"
	"io"
	"log/slog"
	"net/http"

	"github.com/krateoplatformops/mlflow-rest-dynamic-controller-plugin/internal/handlers"
	"github.com/krateoplatformops/mlflow-rest-dynamic-controller-plugin/internal/handlers/modelversion"
)

func CreateModelVersion(opts handlers.HandlerOptions) handlers.Handler {
	return &handler{
		HandlerOptions: opts,
	}
}

var _ handlers.Handler = &handler{}

type handler struct {
	handlers.HandlerOptions
}

// @Summary Create a new model version
// @Description Create a new model version
// @ID create-model-version
// @Accept json
// @Produce json
// @Param name body string true "Register model under this name"
// @Param source body string true "URI indicating the location of the model artifacts"
// @Param run_id body string false "MLflow run ID for correlation, if source was generated by an experiment run in MLflow tracking server"
// @Param tags body []ModelVersionTag false "Additional metadata for model version"
// @Param run_link body string false "MLflow run link - this is the exact link of the run that generated this model version, potentially hosted at another instance of MLflow"
// @Param description body string false "Optional description for model version"
// @Success 200 {object} ModelVersion
// @Router /2.0/mlflow/model-versions/create [post]
func (h *handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	name := r.URL.Query().Get("name")
	source := r.URL.Query().Get("source")
	runID := r.URL.Query().Get("run_id")
	tags := r.URL.Query().Get("tags")
	runLink := r.URL.Query().Get("run_link")
	description := r.URL.Query().Get("description")

	// Validate required parameters
	if name == "" || source == "" {
		http.Error(w, "missing required parameters: name and source", http.StatusBadRequest)
		return
	}

	log := h.Log.With(
		"operation", "/2.0/mlflow/model-versions/create",
		"method", "POST",
		"name", name,
		"source", source,
		"run_id", runID,
		"tags", tags,
		"run_link", runLink,
		"description", description,
	)

	log.Debug("calling MLFlow Model Version API")

	url := h.Server.String() + "/2.0/mlflow/model-versions/create"

	req, err := http.NewRequest("POST", url, r.Body)
	if err != nil {
		h.Log.Error("creating request", "error", err)
		http.Error(w, "failed to create request", http.StatusInternalServerError)
		return
	}

	if r.Header.Get("Authorization") != "" {
		req.Header.Set("Authorization", r.Header.Get("Authorization"))
	}

	resp, err := h.Client.Do(req)
	if err != nil {
		h.Log.Error("calling MLFlow Model Version API", "error", err)
		http.Error(w, "failed to call MLFlow Model Version API", http.StatusInternalServerError)
		return
	}
	defer resp.Body.Close()

	// Read response body
	body, err := io.ReadAll(resp.Body)
	if err != nil {
		log.Error("reading response body", slog.Any("error", err))
		http.Error(w, "failed to read response", http.StatusInternalServerError)
		return
	}

	// Handle non-200 responses
	if resp.StatusCode != http.StatusOK {
		log.Error("MLFlow API error",
			"status", resp.StatusCode,
			"response", string(body),
		)
		http.Error(w, fmt.Sprintf("MLFlow API error: %s", string(body)), resp.StatusCode)
		return
	}

	var model modelversion.ModelVersionResponse

	if err := json.Unmarshal(body, &model); err != nil {
		log.Error("unmarshalling response", slog.Any("error", err))
		http.Error(w, "failed to parse response", http.StatusInternalServerError)
		return
	}

	// Set response headers
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	if err := json.NewEncoder(w).Encode(model.ModelVersion); err != nil {
		log.Error("encoding response", slog.Any("error", err))
		// Cannot write error to client at this point as headers are already sent
		return
	}

	log.Debug("successfully processed request",
		"url", req.URL.String(),
		"status", http.StatusOK,
	)
}